+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 5.0.0
 
 - Output method was changed. Instead of directly returning the results, the process execution is now divided into two steps, spawning and waiting till data ends. So first you need to await on `exec(param, parameters, options)`, which returns an object with the spawned process, and an `output` promise which you can await on again to get the real output.
+- Instead of throwing a normal exit-code error, it now says `Process was killed` when it is
 
 ## 4.0.0